chore: move serving store to under sink
