+++ rust/numaflow-core/src/pipeline.rs
@@ -18,11 +18,11 @@ use crate::pipeline::forwarder::source_forwarder;
 use crate::pipeline::isb::jetstream::reader::JetStreamReader;
 use crate::pipeline::isb::jetstream::writer::JetstreamWriter;
 use crate::pipeline::pipeline::isb::BufferReaderConfig;
-use crate::serving_store::ServingStore;
-use crate::serving_store::nats::NatsServingStore;
-use crate::serving_store::user_defined::UserDefinedStore;
 use crate::shared::create_components;
 use crate::shared::metrics::start_metrics_server;
+use crate::sink::serving_store::ServingStore;
+use crate::sink::serving_store::nats::NatsServingStore;
+use crate::sink::serving_store::user_defined::UserDefinedStore;
 use crate::tracker::TrackerHandle;
 use crate::watermark::WatermarkHandle;
 use crate::watermark::isb::ISBWatermarkHandle;
