+++ rust/numaflow-core/src/shared/create_components.rs
@@ -14,9 +14,9 @@ use crate::config::pipeline::map::{MapMode, MapType, MapVtxConfig};
 use crate::config::pipeline::{DEFAULT_BATCH_MAP_SOCKET, DEFAULT_STREAM_MAP_SOCKET};
 use crate::error::Error;
 use crate::mapper::map::MapHandle;
-use crate::serving_store::ServingStore;
 use crate::shared::grpc;
 use crate::shared::server_info::{ContainerType, sdk_server_info};
+use crate::sink::serving_store::ServingStore;
 use crate::sink::{SinkClientType, SinkWriter, SinkWriterBuilder};
 use crate::source::Source;
 use crate::source::generator::new_generator;
