+++ rust/numaflow-core/src/sink/serving_store/nats.rs
@@ -1,5 +1,5 @@
 use crate::config::pipeline::NatsStoreConfig;
-use crate::serving_store::StoreEntry;
+use crate::sink::serving_store::StoreEntry;
 use async_nats::jetstream::Context;
 use async_nats::jetstream::kv::Store;
 use chrono::Utc;
